Fixed issue with escape key in tmux/byobu
On several computers recently (all 16.04) I've had issues with my backtick
escape key in Byobu. I tried different versions of byobu and the error didn't go
away. So I'm assuming it is an issue related to an updated tmux (2.1 by default
on 16.04; 1.8 on my most commonly used 14.04 system). Haven't quite tracked down
exactly what the issue is, but it is definitely related to byobu setting F12 to
be the prefix key. In my old config I used ` as the prefix key and manually set
F12 to the secondary prefix key to preserve byobu's keybindings. Somehow on the
problem systems this setting wasn't working -- pressing ` simply issued a ` when
in byobu (in tmux the setting worked as anticipated). The workaround was to set
` as the secondary prefix key (prefix2), and then modify the send-prefix
keybinding to send the secondary prefix instead of the prefix by using the '-2'
argument. Now I get the same behavior in tmux and byobu on all of my systems.