Updated tmux send to windows/panes functionality
+ added a new alias that lets me send to all panes and all windows from Bash,
and an alias for sourcing the bashrc
+ The tmux kbd shortcut no longer asks for a session, and it sends to all panes
and windows
+ Removed my modification to emacs incremental search. Was behaving weird for a
long time, and I didn't even realize it. It's a pain to have to hit Backspace
enough times to backup to the start of the search before it starts deleting
characters, but in my method hitting backspace would quickly end up with the
previous search string. This was annoying! Note pressing 'C-g' will delete any
non-matching characters (if you type extra chars and there are no search
matches).

https://stackoverflow.com/questions/16325449/how-to-send-a-command-to-all-panes-in-tmux
https://scripter.co/command-to-every-pane-window-session-in-tmux/