Added a fallback keyboard shortcut for execute-extended-command
Sometimes it is useful to still be able to use the original M-x
behavior (execute-extended-command) -- I often encounter this with
multiple-cursors. Normally, my fallback for this has been to use the "Menu" key
on keyboards that have one. However, on the keyboard I'm using these days, I
don't have that key so I was SOL. Not sure if this shortcut is the best, but
it's easy to remember and I'm unaware of any conflicts